+++ src/main/java/com/financialmeet/DataInitializer.java
@@ -3,7 +3,7 @@ package com.financialmeet;
 import com.financialmeet.model.Account;
 import com.financialmeet.repository.AccountRepository;
 import java.util.ArrayList;
-import java.util.Arrays;
+import java.util.Collections;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.security.crypto.password.PasswordEncoder;
@@ -12,9 +12,11 @@ import org.springframework.stereotype.Component;
 @Component
 public class DataInitializer implements CommandLineRunner {
 
-  @Autowired AccountRepository accountRepository;
+  @Autowired
+  private AccountRepository accountRepository;
 
-  @Autowired PasswordEncoder passwordEncoder;
+  @Autowired
+  private PasswordEncoder passwordEncoder;
 
   @Override
   public void run(String... args) {
@@ -23,13 +25,10 @@ public class DataInitializer implements CommandLineRunner {
     Account user = new Account();
     user.setUsername("user");
     user.setPassword(this.passwordEncoder.encode("password"));
-    user.setRoles(Arrays.asList("ROLE_USER"));
+    user.setRoles(Collections.singletonList("ROLE_USER"));
 
     accountRepository.save(user);
 
-
   }
-
-
 
 }
