+++ src/main/java/com/financialmeet/config/SecurityConfig.java
@@ -36,4 +36,5 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
         .and()
         .apply(new JwtConfigurer(jwtTokenProvider));
   }
+
 }
