+++ src/main/resources/application.properties
@@ -1,5 +1,5 @@
 spring.jpa.hibernate.ddl-auto=create
-spring.datasource.url=YOUR_DB_URL
-spring.datasource.username=USERNAME
-spring.datasource.password=PASSWORD
+spring.datasource.url=YOUR_URL
+spring.datasource.username=username
+spring.datasource.password=password
 spring.jackson.serialization.fail-on-empty-beans=false
